Get role put into SearchVC
